requirements.txt
Django~=5.0.4
Pillow
flake8
django-filter~=24.2
djangorestframework~=3.15.1